Put web3-agent as main chat. Remove the others
