+++ app/normal_chat/page.tsx
@@ -1,0 +1,63 @@
+import { ChatWindow } from "@/components/ChatWindow";
+import { GuideInfoBox } from "@/components/guide/GuideInfoBox";
+
+export default function Home() {
+  const InfoCard = (
+    <GuideInfoBox>
+      <ul>
+        <li className="text-l">
+          🤝
+          <span className="ml-2">
+            This template showcases a simple chatbot using{" "}
+            <a href="https://js.langchain.com/" target="_blank">
+              LangChain.js
+            </a>{" "}
+            and the Vercel{" "}
+            <a href="https://sdk.vercel.ai/docs" target="_blank">
+              AI SDK
+            </a>{" "}
+            in a{" "}
+            <a href="https://nextjs.org/" target="_blank">
+              Next.js
+            </a>{" "}
+            project.
+          </span>
+        </li>
+        <li className="hidden text-l md:block">
+          💻
+          <span className="ml-2">
+            You can find the prompt and model logic for this use-case in{" "}
+            <code>app/api/chat/route.ts</code>.
+          </span>
+        </li>
+        <li>
+          🏴‍☠️
+          <span className="ml-2">
+            By default, the bot is pretending to be a pirate, but you can change
+            the prompt to whatever you want!
+          </span>
+        </li>
+        <li className="hidden text-l md:block">
+          🎨
+          <span className="ml-2">
+            The main frontend logic is found in <code>app/page.tsx</code>.
+          </span>
+        </li>
+        <li className="text-l">
+          👇
+          <span className="ml-2">
+            Try asking e.g. <code>What is it like to be a pirate?</code> below!
+          </span>
+        </li>
+      </ul>
+    </GuideInfoBox>
+  );
+  return (
+    <ChatWindow
+      endpoint="api/chat"
+      emoji="🏴‍☠️"
+      placeholder="I'm an LLM pretending to be a pirate! Ask me about the pirate life!"
+      emptyStateComponent={InfoCard}
+    />
+  );
+}
